Refactor camera handling and improve code clarity
- Updated event class definitions in `ApplicationEvent.h`.
- Changed `KeyCode` type from `uint8_t` to `uint16_t` in `KeyCodes.h`.
- Added `<sstream>` include in `KeyEvent.h`.
- Refactored `Camera` constructor and methods in `Camera.cpp` and `Camera.h` to use `CameraController` directly.
- Modified movement function signatures in `CameraController` to allow state modification.
- Changed skybox member variables from pointers to unsigned integers in `Renderer.h`.
- Added `[[nodiscard]]` attributes to getter methods in `Texture.h`.
- Simplified `CameraReset` calls in `ImGuiManager.cpp`.
- Clarified comments regarding the `camera` pointer in `ImGuiManager.h`.
- Renamed logging macros in `Logging.h` for consistency.
- Added TODO comments in `main.cpp` for future improvements.

+++ SpectralForge-Core/KeyCodes.h
@@ -2,7 +2,7 @@
 
 namespace Badiya
 {
-	using KeyCode = uint8_t;
+	using KeyCode = uint16_t;
 
 	namespace Key
 	{
